CreatingMoleculeClass - FINISHED
Molecule Class modified based on work done in
ValidatingAtomMoleculeInput branch. MoleculeTester class created (do not
include in final submission), and everything works as expected. NOTE:
there is currently no validation for if all the atoms in a molecule
exist. This class must be used on valid molecules. Still need to finish
javadoc descriptions.