Switching between windows
Calculate_Reaction_TextFields class modified to have instance variables
related to the user input arrays and error message array. Reaction
Controller class modified to a)have a constructor instead of the old
primary method, and this constructor creates a new stage; b)two methods
to handle the action of the enter button depending on the validity of
the user input; c)has a showStage method. The Project View COntroller
was modified to invoke a new Reaction_View_COntroller and call its
showStage method